test_requirements.txt
django-load
django-nose
coverage
django-coverage
django-jasmine
ipdb
flake8
fabric
factory_boy
mock
selenium
south
